main

+++ works/10.8/work2.cpp
@@ -2,10 +2,10 @@
 int main()
 {
     using namespace std;
-    int r ;
+    double r ;
     cout <<"r=";   
     cin >> r;
-    int area=r*r*3.14;
-    int circum=2*r*3.14;
+    double area=r*r*3.14;
+    double circum=2*r*3.14;
     cout<<"area="<<area<<"  circum="<<circum<<endl;
 }
